move code into file structure as requested
