+++ internal/service/artifact_handler.go
@@ -1,0 +1,9 @@
+package service
+
+import (
+	"gorm.io/gorm"
+)
+
+type artifactHandler struct {
+	db *gorm.DB
+}
