+++ internal/service/db_service_tx.go
@@ -1,0 +1,27 @@
+package service
+
+import (
+	"context"
+
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
+
+	"github.com/opendatahub-io/model-registry/internal/server"
+)
+
+func handleTransaction(ctx context.Context, err *error) {
+	// handle panic
+	if perr := recover(); perr != nil {
+		_ = server.Rollback(ctx)
+		*err = status.Errorf(codes.Internal, "server panic: %v", perr)
+		return
+	}
+	if err == nil || *err == nil {
+		*err = server.Commit(ctx)
+	} else {
+		_ = server.Rollback(ctx)
+		if _, ok := status.FromError(*err); !ok {
+			*err = status.Errorf(codes.Internal, "internal error: %v", *err)
+		}
+	}
+}
